Issue #7116: remove AbstractTypeAwareCheck and combine to JavadocMethod
