+++ .ci/pitest.sh
@@ -191,13 +191,11 @@ pitest-javadoc)
   "AbstractJavadocCheck.java.html:<td class='covered'><pre><span  class='survived'>        beginJavadocTree(root);</span></pre></td></tr>"
   "AbstractJavadocCheck.java.html:<td class='covered'><pre><span  class='survived'>        finishJavadocTree(root);</span></pre></td></tr>"
   "AbstractJavadocCheck.java.html:<td class='covered'><pre><span  class='survived'>        javadocTokens.clear();</span></pre></td></tr>"
-  "AbstractTypeAwareCheck.java.html:<td class='covered'><pre><span  class='survived'>        if (!currentClassName.isEmpty()) {</span></pre></td></tr>"
-  "AbstractTypeAwareCheck.java.html:<td class='covered'><pre><span  class='survived'>            if (dotIdx == -1) {</span></pre></td></tr>"
-  "AbstractTypeAwareCheck.java.html:<td class='covered'><pre><span  class='survived'>        imports.clear();</span></pre></td></tr>"
-  "AbstractTypeAwareCheck.java.html:<td class='covered'><pre><span  class='survived'>        typeParams.clear();</span></pre></td></tr>"
-  "AbstractTypeAwareCheck.java.html:<td class='covered'><pre><span  class='survived'>            loadable = clazz != null;</span></pre></td></tr>"
-  "AbstractTypeAwareCheck.java.html:<td class='covered'><pre><span  class='survived'>            logLoadError(ident);</span></pre></td></tr>"
-  "AbstractTypeAwareCheck.java.html:<td class='covered'><pre><span  class='survived'>        if (clazz == null) {</span></pre></td></tr>"
+  "JavadocMethodCheck.java.html:<td class='covered'><pre><span  class='survived'>        if (!currentClassName.isEmpty()) {</span></pre></td></tr>"
+  "JavadocMethodCheck.java.html:<td class='covered'><pre><span  class='survived'>            if (dotIdx == -1) {</span></pre></td></tr>"
+  "JavadocMethodCheck.java.html:<td class='covered'><pre><span  class='survived'>        imports.clear();</span></pre></td></tr>"
+  "JavadocMethodCheck.java.html:<td class='covered'><pre><span  class='survived'>        currentTypeParams.clear();</span></pre></td></tr>"
+  "JavadocMethodCheck.java.html:<td class='covered'><pre><span  class='survived'>            loadable = clazz != null;</span></pre></td></tr>"
   "ClassResolver.java.html:<td class='uncovered'><pre><span  class=''>        catch (final ClassNotFoundException ex) {</span></pre></td></tr>"
   "ClassResolver.java.html:<td class='uncovered'><pre><span  class='survived'>            throw new IllegalStateException(ex);</span></pre></td></tr>"
   "JavadocTagInfo.java.html:<td class='covered'><pre><span  class='survived'>            .collect(Collectors.toMap(JavadocTagInfo::getName, tagName -&#62; tagName)));</span></pre></td></tr>"
