+++ src/main/java/com/puppycrawl/tools/checkstyle/checks/javadoc/JavadocMethodCheck.java
@@ -19,17 +19,23 @@
 
 package com.puppycrawl.tools.checkstyle.checks.javadoc;
 
+import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.Deque;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.ListIterator;
+import java.util.Map;
 import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+import com.puppycrawl.tools.checkstyle.FileStatefulCheck;
+import com.puppycrawl.tools.checkstyle.api.AbstractCheck;
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
 import com.puppycrawl.tools.checkstyle.api.FileContents;
 import com.puppycrawl.tools.checkstyle.api.FullIdent;
@@ -43,11 +49,9 @@ import com.puppycrawl.tools.checkstyle.utils.ScopeUtil;
 
 /**
  * Checks the Javadoc of a method or constructor.
- *
- *
- * @noinspection deprecation
  */
-public class JavadocMethodCheck extends AbstractTypeAwareCheck {
+@FileStatefulCheck
+public class JavadocMethodCheck extends AbstractCheck {
 
     /**
      * A key is pointing to the warning message text in "messages.properties"
@@ -117,6 +121,21 @@ public class JavadocMethodCheck extends AbstractTypeAwareCheck {
     /** Compiled regexp to match Javadoc tags with no argument and {}. */
     private static final Pattern MATCH_JAVADOC_NOARG_CURLY =
             CommonUtil.createPattern("\\{\\s*@(inheritDoc)\\s*\\}");
+
+    /** Stack of maps for type params. */
+    private final Deque<Map<String, AbstractClassInfo>> currentTypeParams = new ArrayDeque<>();
+
+    /** Imports details. **/
+    private final Set<String> imports = new HashSet<>();
+
+    /** Full identifier for package of the method. **/
+    private FullIdent packageFullIdent;
+
+    /** Name of current class. */
+    private String currentClassName;
+
+    /** {@code ClassResolver} instance for current tree. */
+    private ClassResolver classResolver;
 
     /** The visibility scope where Javadoc comments are checked. */
     private Scope scope = Scope.PRIVATE;
@@ -252,6 +271,40 @@ public class JavadocMethodCheck extends AbstractTypeAwareCheck {
         allowMissingReturnTag = flag;
     }
 
+    /**
+     * Controls whether to log class loading errors to the checkstyle report
+     * instead of throwing a RTE.
+     *
+     * @param logLoadErrors true if errors should be logged
+     * @deprecated No substitute.
+     */
+    @Deprecated
+    public final void setLogLoadErrors(boolean logLoadErrors) {
+        // no code
+    }
+
+    /**
+     * Controls whether to show class loading errors in the checkstyle report.
+     *
+     * @param suppressLoadErrors true if errors shouldn't be shown
+     * @deprecated No substitute.
+     */
+    @Deprecated
+    public final void setSuppressLoadErrors(boolean suppressLoadErrors) {
+        // no code
+    }
+
+    @Override
+    public final int[] getRequiredTokens() {
+        return new int[] {
+            TokenTypes.PACKAGE_DEF,
+            TokenTypes.IMPORT,
+            TokenTypes.CLASS_DEF,
+            TokenTypes.INTERFACE_DEF,
+            TokenTypes.ENUM_DEF,
+        };
+    }
+
     @Override
     public int[] getDefaultTokens() {
         return getAcceptableTokens();
@@ -272,7 +325,68 @@ public class JavadocMethodCheck extends AbstractTypeAwareCheck {
     }
 
     @Override
-    protected final void processAST(DetailAST ast) {
+    public void beginTree(DetailAST rootAST) {
+        packageFullIdent = FullIdent.createFullIdent(null);
+        imports.clear();
+        // add java.lang.* since it's always imported
+        imports.add("java.lang.*");
+        classResolver = null;
+        currentClassName = "";
+        currentTypeParams.clear();
+    }
+
+    @Override
+    public final void visitToken(DetailAST ast) {
+        if (ast.getType() == TokenTypes.PACKAGE_DEF) {
+            processPackage(ast);
+        }
+        else if (ast.getType() == TokenTypes.IMPORT) {
+            processImport(ast);
+        }
+        else if (ast.getType() == TokenTypes.CLASS_DEF
+                 || ast.getType() == TokenTypes.INTERFACE_DEF
+                 || ast.getType() == TokenTypes.ENUM_DEF) {
+            processClass(ast);
+        }
+        else {
+            if (ast.getType() == TokenTypes.METHOD_DEF) {
+                processTypeParams(ast);
+            }
+            processAST(ast);
+        }
+    }
+
+    @Override
+    public final void leaveToken(DetailAST ast) {
+        if (ast.getType() == TokenTypes.CLASS_DEF
+            || ast.getType() == TokenTypes.INTERFACE_DEF
+            || ast.getType() == TokenTypes.ENUM_DEF) {
+            // perhaps it was inner class
+            int dotIdx = currentClassName.lastIndexOf('$');
+            if (dotIdx == -1) {
+                // perhaps just a class
+                dotIdx = currentClassName.lastIndexOf('.');
+            }
+            if (dotIdx == -1) {
+                // looks like a topmost class
+                currentClassName = "";
+            }
+            else {
+                currentClassName = currentClassName.substring(0, dotIdx);
+            }
+            currentTypeParams.pop();
+        }
+        else if (ast.getType() == TokenTypes.METHOD_DEF) {
+            currentTypeParams.pop();
+        }
+    }
+
+    /**
+     * Called to process an AST when visiting it.
+     * @param ast the AST to process. Guaranteed to not be PACKAGE_DEF or
+     *             IMPORT tokens.
+     */
+    private void processAST(DetailAST ast) {
         final Scope theScope = calculateScope(ast);
         if (shouldCheck(ast, theScope)) {
             final FileContents contents = getFileContents();
@@ -282,11 +396,6 @@ public class JavadocMethodCheck extends AbstractTypeAwareCheck {
                 checkComment(ast, textBlock);
             }
         }
-    }
-
-    @Override
-    protected final void logLoadError(Token ident) {
-        // no code because class loading is deprecated
     }
 
     /**
@@ -553,7 +662,7 @@ public class JavadocMethodCheck extends AbstractTypeAwareCheck {
                         || child.getType() == TokenTypes.DOT) {
                     final FullIdent ident = FullIdent.createFullIdent(child);
                     final ExceptionInfo exceptionInfo = new ExceptionInfo(
-                            createClassInfo(new Token(ident), getCurrentClassName()));
+                            createClassInfo(new Token(ident), currentClassName));
                     returnValue.add(exceptionInfo);
                 }
                 child = child.getNextSibling();
@@ -726,7 +835,7 @@ public class JavadocMethodCheck extends AbstractTypeAwareCheck {
             final Token token = new Token(tag.getFirstArg(), tag.getLineNo(), tag
                     .getColumnNo());
             final AbstractClassInfo documentedClassInfo = createClassInfo(token,
-                    getCurrentClassName());
+                    currentClassName);
             final boolean found = foundThrows.contains(documentedEx)
                     || isInThrows(throwsList, documentedClassInfo, foundThrows);
 
@@ -800,6 +909,372 @@ public class JavadocMethodCheck extends AbstractTypeAwareCheck {
         }
 
         return found;
+    }
+
+    /**
+     * Is exception is unchecked (subclass of {@code RuntimeException}
+     * or {@code Error}.
+     *
+     * @param exception {@code Class} of exception to check
+     * @return true  if exception is unchecked
+     *         false if exception is checked
+     */
+    private static boolean isUnchecked(Class<?> exception) {
+        return isSubclass(exception, RuntimeException.class)
+            || isSubclass(exception, Error.class);
+    }
+
+    /**
+     * Checks if one class is subclass of another.
+     *
+     * @param child {@code Class} of class
+     *               which should be child
+     * @param parent {@code Class} of class
+     *                which should be parent
+     * @return true  if aChild is subclass of aParent
+     *         false otherwise
+     */
+    private static boolean isSubclass(Class<?> child, Class<?> parent) {
+        return parent != null && child != null
+            && parent.isAssignableFrom(child);
+    }
+
+    /**
+     * Returns the current tree's ClassResolver.
+     * @return {@code ClassResolver} for current tree.
+     */
+    private ClassResolver getClassResolver() {
+        if (classResolver == null) {
+            classResolver =
+                new ClassResolver(getClass().getClassLoader(),
+                                  packageFullIdent.getText(),
+                                  imports);
+        }
+        return classResolver;
+    }
+
+    /**
+     * Attempts to resolve the Class for a specified name.
+     * @param resolvableClassName name of the class to resolve
+     * @param className name of surrounding class.
+     * @return the resolved class or {@code null}
+     *          if unable to resolve the class.
+     * @noinspection WeakerAccess
+     */
+    // -@cs[ForbidWildcardAsReturnType] The class is deprecated and will be removed soon.
+    private Class<?> resolveClass(String resolvableClassName,
+                                          String className) {
+        Class<?> clazz;
+        try {
+            clazz = getClassResolver().resolve(resolvableClassName, className);
+        }
+        // -@cs[IllegalCatch] Exception type is not predictable.
+        catch (final Exception ignored) {
+            clazz = null;
+        }
+        return clazz;
+    }
+
+    /**
+     * Tries to load class. Logs error if unable.
+     * @param ident name of class which we try to load.
+     * @param className name of surrounding class.
+     * @return {@code Class} for a ident.
+     * @noinspection WeakerAccess, MethodOnlyUsedFromInnerClass
+     */
+    // -@cs[ForbidWildcardAsReturnType] The class is deprecated and will be removed soon.
+    private Class<?> tryLoadClass(Token ident, String className) {
+        return resolveClass(ident.getText(), className);
+    }
+
+    /**
+     * Collects the details of a package.
+     * @param ast node containing the package details
+     */
+    private void processPackage(DetailAST ast) {
+        final DetailAST nameAST = ast.getLastChild().getPreviousSibling();
+        packageFullIdent = FullIdent.createFullIdent(nameAST);
+    }
+
+    /**
+     * Collects the details of imports.
+     * @param ast node containing the import details
+     */
+    private void processImport(DetailAST ast) {
+        final FullIdent name = FullIdent.createFullIdentBelow(ast);
+        imports.add(name.getText());
+    }
+
+    /**
+     * Process type params (if any) for given class, enum or method.
+     * @param ast class, enum or method to process.
+     */
+    private void processTypeParams(DetailAST ast) {
+        final DetailAST params =
+            ast.findFirstToken(TokenTypes.TYPE_PARAMETERS);
+
+        final Map<String, AbstractClassInfo> paramsMap = new HashMap<>();
+        currentTypeParams.push(paramsMap);
+
+        if (params != null) {
+            for (DetailAST child = params.getFirstChild();
+                 child != null;
+                 child = child.getNextSibling()) {
+                if (child.getType() == TokenTypes.TYPE_PARAMETER) {
+                    final DetailAST bounds =
+                        child.findFirstToken(TokenTypes.TYPE_UPPER_BOUNDS);
+                    if (bounds != null) {
+                        final FullIdent name =
+                            FullIdent.createFullIdentBelow(bounds);
+                        final AbstractClassInfo classInfo =
+                            createClassInfo(new Token(name), currentClassName);
+                        final String alias =
+                                child.findFirstToken(TokenTypes.IDENT).getText();
+                        paramsMap.put(alias, classInfo);
+                    }
+                }
+            }
+        }
+    }
+
+    /**
+     * Processes class definition.
+     * @param ast class definition to process.
+     */
+    private void processClass(DetailAST ast) {
+        final DetailAST ident = ast.findFirstToken(TokenTypes.IDENT);
+        String innerClass = ident.getText();
+
+        if (!currentClassName.isEmpty()) {
+            innerClass = "$" + innerClass;
+        }
+        currentClassName += innerClass;
+        processTypeParams(ast);
+    }
+
+    /**
+     * Creates class info for given name.
+     * @param name name of type.
+     * @param surroundingClass name of surrounding class.
+     * @return class info for given name.
+     */
+    private AbstractClassInfo createClassInfo(final Token name,
+                                              final String surroundingClass) {
+        final AbstractClassInfo result;
+        final AbstractClassInfo classInfo = findClassAlias(name.getText());
+        if (classInfo == null) {
+            result = new RegularClass(name, surroundingClass, this);
+        }
+        else {
+            result = new ClassAlias(name, classInfo);
+        }
+        return result;
+    }
+
+    /**
+     * Looking if a given name is alias.
+     * @param name given name
+     * @return ClassInfo for alias if it exists, null otherwise
+     * @noinspection WeakerAccess
+     */
+    private AbstractClassInfo findClassAlias(final String name) {
+        AbstractClassInfo classInfo = null;
+        final Iterator<Map<String, AbstractClassInfo>> iterator = currentTypeParams
+                .descendingIterator();
+        while (iterator.hasNext()) {
+            final Map<String, AbstractClassInfo> paramMap = iterator.next();
+            classInfo = paramMap.get(name);
+            if (classInfo != null) {
+                break;
+            }
+        }
+        return classInfo;
+    }
+
+    /**
+     * Contains class's {@code Token}.
+     */
+    private abstract static class AbstractClassInfo {
+
+        /** {@code FullIdent} associated with this class. */
+        private final Token name;
+
+        /**
+         * Creates new instance of class information object.
+         * @param className token which represents class name.
+         */
+        protected AbstractClassInfo(final Token className) {
+            if (className == null) {
+                throw new IllegalArgumentException(
+                    "ClassInfo's name should be non-null");
+            }
+            name = className;
+        }
+
+        /**
+         * Returns class associated with that object.
+         * @return {@code Class} associated with an object.
+         */
+        // -@cs[ForbidWildcardAsReturnType] The class is deprecated and will be removed soon.
+        public abstract Class<?> getClazz();
+
+        /**
+         * Gets class name.
+         * @return class name
+         */
+        public final Token getName() {
+            return name;
+        }
+
+    }
+
+    /** Represents regular classes/enums. */
+    private static final class RegularClass extends AbstractClassInfo {
+
+        /** Name of surrounding class. */
+        private final String surroundingClass;
+        /** The check we use to resolve classes. */
+        private final JavadocMethodCheck check;
+        /** Is class loadable. */
+        private boolean loadable = true;
+        /** {@code Class} object of this class if it's loadable. */
+        private Class<?> classObj;
+
+        /**
+         * Creates new instance of of class information object.
+         * @param name {@code FullIdent} associated with new object.
+         * @param surroundingClass name of current surrounding class.
+         * @param check the check we use to load class.
+         */
+        /* package */ RegularClass(final Token name,
+                             final String surroundingClass,
+                             final JavadocMethodCheck check) {
+            super(name);
+            this.surroundingClass = surroundingClass;
+            this.check = check;
+        }
+
+        @Override
+        public Class<?> getClazz() {
+            if (loadable && classObj == null) {
+                setClazz(check.tryLoadClass(getName(), surroundingClass));
+            }
+            return classObj;
+        }
+
+        /**
+         * Associates {@code Class} with an object.
+         * @param clazz {@code Class} to associate with.
+         */
+        private void setClazz(Class<?> clazz) {
+            classObj = clazz;
+            loadable = clazz != null;
+        }
+
+        @Override
+        public String toString() {
+            return "RegularClass[name=" + getName()
+                    + ", in class='" + surroundingClass + '\''
+                    + ", check=" + check.hashCode()
+                    + ", loadable=" + loadable
+                    + ", class=" + classObj
+                    + ']';
+        }
+
+    }
+
+    /** Represents type param which is "alias" for real type. */
+    private static class ClassAlias extends AbstractClassInfo {
+
+        /** Class information associated with the alias. */
+        private final AbstractClassInfo classInfo;
+
+        /**
+         * Creates new instance of the class.
+         * @param name token which represents name of class alias.
+         * @param classInfo class information associated with the alias.
+         */
+        /* package */ ClassAlias(final Token name, AbstractClassInfo classInfo) {
+            super(name);
+            this.classInfo = classInfo;
+        }
+
+        @Override
+        public final Class<?> getClazz() {
+            return classInfo.getClazz();
+        }
+
+        @Override
+        public String toString() {
+            return "ClassAlias[alias " + getName() + " for " + classInfo.getName() + "]";
+        }
+
+    }
+
+    /**
+     * Represents text element with location in the text.
+     */
+    private static class Token {
+
+        /** Token's column number. */
+        private final int columnNo;
+        /** Token's line number. */
+        private final int lineNo;
+        /** Token's text. */
+        private final String text;
+
+        /**
+         * Creates token.
+         * @param text token's text
+         * @param lineNo token's line number
+         * @param columnNo token's column number
+         */
+        /* default */ Token(String text, int lineNo, int columnNo) {
+            this.text = text;
+            this.lineNo = lineNo;
+            this.columnNo = columnNo;
+        }
+
+        /**
+         * Converts FullIdent to Token.
+         * @param fullIdent full ident to convert.
+         */
+        /* default */ Token(FullIdent fullIdent) {
+            text = fullIdent.getText();
+            lineNo = fullIdent.getLineNo();
+            columnNo = fullIdent.getColumnNo();
+        }
+
+        /**
+         * Gets line number of the token.
+         * @return line number of the token
+         */
+        public int getLineNo() {
+            return lineNo;
+        }
+
+        /**
+         * Gets column number of the token.
+         * @return column number of the token
+         */
+        public int getColumnNo() {
+            return columnNo;
+        }
+
+        /**
+         * Gets text of the token.
+         * @return text of the token
+         */
+        public String getText() {
+            return text;
+        }
+
+        @Override
+        public String toString() {
+            return "Token[" + text + "(" + lineNo
+                + "x" + columnNo + ")]";
+        }
+
     }
 
     /** Stores useful information about declared exception. */
