+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/javadoc/JavadocMethodCheckTest.java
@@ -26,6 +26,15 @@ import static com.puppycrawl.tools.checkstyle.checks.javadoc.JavadocMethodCheck.
 import static com.puppycrawl.tools.checkstyle.checks.javadoc.JavadocMethodCheck.MSG_UNUSED_TAG;
 import static com.puppycrawl.tools.checkstyle.checks.javadoc.JavadocMethodCheck.MSG_UNUSED_TAG_GENERAL;
 import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
+
+import java.lang.reflect.Constructor;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
 
 import org.junit.Test;
 
@@ -479,6 +488,146 @@ public class JavadocMethodCheckTest extends AbstractModuleTestSupport {
             "10:50: " + getCheckMessage(MSG_EXPECTED_TAG, "@param", "p1"),
         };
         verify(checkConfig, getPath("InputJavadocMethodConstructor.java"), expected);
+    }
+
+    @Test
+    public void testIsSubclassWithNulls() throws Exception {
+        final JavadocMethodCheck check = new JavadocMethodCheck();
+        final Method method = check.getClass()
+                .getDeclaredMethod("isSubclass", Class.class, Class.class);
+        method.setAccessible(true);
+        assertFalse("Should return false if at least one of the params is null",
+            (boolean) method.invoke(check, null, null));
+    }
+
+    @Test
+    public void testGetRequiredTokens() {
+        final int[] expected = {
+            TokenTypes.PACKAGE_DEF,
+            TokenTypes.IMPORT,
+            TokenTypes.CLASS_DEF,
+            TokenTypes.INTERFACE_DEF,
+            TokenTypes.ENUM_DEF,
+        };
+        final JavadocMethodCheck check = new JavadocMethodCheck();
+        final int[] actual = check.getRequiredTokens();
+        assertArrayEquals("Required tokens differ from expected", expected, actual);
+    }
+
+    @Test
+    public void testTokenToString() throws Exception {
+        final Class<?> tokenType = Class.forName("com.puppycrawl.tools.checkstyle.checks.javadoc."
+                + "JavadocMethodCheck$Token");
+        final Constructor<?> tokenConstructor = tokenType.getDeclaredConstructor(String.class,
+                int.class, int.class);
+        final Object token = tokenConstructor.newInstance("blablabla", 1, 1);
+        final Method toString = token.getClass().getDeclaredMethod("toString");
+        final String result = (String) toString.invoke(token);
+        assertEquals("Invalid toString result", "Token[blablabla(1x1)]", result);
+    }
+
+    @Test
+    public void testClassRegularClass() throws Exception {
+        final Class<?> tokenType = Class.forName("com.puppycrawl.tools.checkstyle.checks.javadoc."
+                + "JavadocMethodCheck$Token");
+
+        final Class<?> regularClassType = Class
+                .forName("com.puppycrawl.tools.checkstyle.checks.javadoc."
+                        + "JavadocMethodCheck$RegularClass");
+        final Constructor<?> regularClassConstructor = regularClassType.getDeclaredConstructor(
+                tokenType, String.class, JavadocMethodCheck.class);
+        regularClassConstructor.setAccessible(true);
+
+        try {
+            regularClassConstructor.newInstance(null, "", new JavadocMethodCheck());
+            fail("Exception is expected");
+        }
+        catch (InvocationTargetException ex) {
+            assertTrue("Invalid exception class, expected: IllegalArgumentException.class",
+                ex.getCause() instanceof IllegalArgumentException);
+            assertEquals("Invalid exception message",
+                "ClassInfo's name should be non-null", ex.getCause().getMessage());
+        }
+
+        final Constructor<?> tokenConstructor = tokenType.getDeclaredConstructor(String.class,
+                int.class, int.class);
+        final Object token = tokenConstructor.newInstance("blablabla", 1, 1);
+
+        final JavadocMethodCheck methodCheck = new JavadocMethodCheck();
+        final Object regularClass = regularClassConstructor.newInstance(token, "sur",
+                methodCheck);
+
+        final Method toString = regularClass.getClass().getDeclaredMethod("toString");
+        toString.setAccessible(true);
+        final String result = (String) toString.invoke(regularClass);
+        final String expected = "RegularClass[name=Token[blablabla(1x1)], in class='sur', check="
+                + methodCheck.hashCode() + "," + " loadable=true, class=null]";
+
+        assertEquals("Invalid toString result", expected, result);
+
+        final Method setClazz = regularClass.getClass().getDeclaredMethod("setClazz", Class.class);
+        setClazz.setAccessible(true);
+        final Class<?> arg = null;
+        setClazz.invoke(regularClass, arg);
+
+        final Method getClazz = regularClass.getClass().getDeclaredMethod("getClazz");
+        getClazz.setAccessible(true);
+        assertNull("Expected null", getClazz.invoke(regularClass));
+    }
+
+    @Test
+    public void testClassAliasToString() throws Exception {
+        final Class<?> tokenType = Class.forName("com.puppycrawl.tools.checkstyle.checks.javadoc."
+                + "JavadocMethodCheck$Token");
+
+        final Class<?> regularClassType = Class
+                .forName("com.puppycrawl.tools.checkstyle.checks.javadoc."
+                        + "JavadocMethodCheck$RegularClass");
+        final Constructor<?> regularClassConstructor = regularClassType.getDeclaredConstructor(
+                tokenType, String.class, JavadocMethodCheck.class);
+        regularClassConstructor.setAccessible(true);
+
+        final Constructor<?> tokenConstructor = tokenType.getDeclaredConstructor(String.class,
+                int.class, int.class);
+        final Object token = tokenConstructor.newInstance("blablabla", 1, 1);
+
+        final Object regularClass = regularClassConstructor.newInstance(token, "sur",
+                new JavadocMethodCheck());
+
+        final Class<?> classAliasType = Class.forName(
+                "com.puppycrawl.tools.checkstyle.checks.javadoc.JavadocMethodCheck$ClassAlias");
+        final Class<?> abstractTypeInfoType = Class
+                .forName("com.puppycrawl.tools.checkstyle.checks.javadoc."
+                        + "JavadocMethodCheck$AbstractClassInfo");
+
+        final Constructor<?> classAliasConstructor = classAliasType
+                .getDeclaredConstructor(tokenType, abstractTypeInfoType);
+        classAliasConstructor.setAccessible(true);
+
+        final Object classAlias = classAliasConstructor.newInstance(token, regularClass);
+        final Method toString = classAlias.getClass().getDeclaredMethod("toString");
+        toString.setAccessible(true);
+        final String result = (String) toString.invoke(classAlias);
+        assertEquals("Invalid toString result",
+            "ClassAlias[alias Token[blablabla(1x1)] for Token[blablabla(1x1)]]", result);
+    }
+
+    @Test
+    public void testWithoutLogErrors() throws Exception {
+        final DefaultConfiguration config = createModuleConfig(JavadocMethodCheck.class);
+        config.addAttribute("allowUndeclaredRTE", "true");
+        verify(config, getPath("InputJavadocMethodLoadErrors.java"),
+                CommonUtil.EMPTY_STRING_ARRAY);
+    }
+
+    @Test
+    public void testWithSuppressLoadErrors() throws Exception {
+        final DefaultConfiguration checkConfig = createModuleConfig(JavadocMethodCheck.class);
+        checkConfig.addAttribute("suppressLoadErrors", "true");
+        checkConfig.addAttribute("allowUndeclaredRTE", "true");
+        final String[] expected = CommonUtil.EMPTY_STRING_ARRAY;
+
+        verify(checkConfig, getPath("InputJavadocMethodLoadErrors.java"), expected);
     }
 
 }
