--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/abstracttypeaware/InputAbstractTypeAwareLoadErrors.java
@@ -1,10 +1,0 @@
-package com.puppycrawl.tools.checkstyle.checks.javadoc.abstracttypeaware;
-
-public class InputAbstractTypeAwareLoadErrors
-{
-    /**
-     * aasdf
-     * @throws InvalidExceptionName exception that cannot be loaded
-     */
-    void method() {}
-}
